fix warning

+++ source/adapter/GUM_DTex.cpp
@@ -190,7 +190,8 @@ draw_flush()
 static void
 scissor_push(int x, int y, int w, int h)
 {
-	s2::RenderScissor::Instance()->Push(x, y, w, h, false, true);
+	s2::RenderScissor::Instance()->Push(
+		static_cast<float>(x), static_cast<float>(y), static_cast<float>(w), static_cast<float>(h), false, true);
 }
 
 static void
